Move response to fixture file, find image by alt
